Remove language implying meritocratic attendance
The HACS website currently states:

> As a yearly event, we’ve been bringing the world’s best cryptographic developers (from open source and industry) together with top experts in formal verification and high assurance methodologies. 

HACS is not a meritocratic event; it is a closed event based on an opaque and unaccountable governance structure and attendance criteria. As such, it should not claim that it brings together "the world's best", "top experts", etc. since that is meritocratic language.

On the side, that language is also just obtuse and pretentious, and the organizers can probably do better.

This edit improves the text on the website to address this issue.